requirements.txt
aiofiles           
aiogram            
APScheduler        
aiomysql            
magic-filter       
python-dotenv      
requests           
urllib3            
